IMPORTANT UPDATE
This version of learning.html and learning.js returns the results as long as they match at least ONE of the checkboxes. This means that partial matches are being returned! Now I need to figure out how to maybe add a percentage or something. Will optimize for website use soon